requirements.txt
python-dotenv==1.1.1
structlog

langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72

langchain-groq==0.3.6
langchain-huggingface==0.1.2
sentence-transformers==3.3.1

docx2txt==0.9
ipykernel==6.30.0
python-multipart==0.0.20
faiss-cpu==1.9.0.post1

fastapi==0.115.6
uvicorn==0.32.1
Jinja2==3.1.4

langsmith
pandas
PyYAML
pypdf
